Off The Clock Refactoring
No big deal, but singpolyma pointed out to me that my ~<- function could
be written in this way instead and I liked it enough to make the change.

I considered for a moment not even keeping ~<- and using:
(lens ~=) =<< func
everywhere instead.

For now I've decided to keep it, but implement it in the above way.

We'll see if it stays.