Added Game Over and Quitting
Changed how handle_event works.

It no longer does the recursion.
Now it returns a Bool, and that's checked.

The reason is that there are other conditions that can lead to the game
leaving the main loop.

For example, now it goes to game_over if thirst is over max.
Right now game_over just quits, though...